New step to bootstrap confidence intervals
Chris and I think rather than using +/- 2.5% on the 30 year mean and
variance that we should use bootstrapped confidence intervals. Python
doesn’t have a good library for bootstrapping, so here we will
calculate bootstrapped values for the 95% confidence level upper and
lower bounds on the mean and variance.

This is an initial draft commit. It doesn’t work yet, but I’m putting
it up so Chris and I can collaborate.