beginning of revision edits
this is the first of a number of commits associated with edits. most of
the files are from the two matlab toolboxes (mcmcstat and dlmtbx).

in addition, I’ve modified the step10 file to include information about
the mean and standard deviation of the dlm level and slope for chris to
use to test some of the underlying assumptions in response to a
reviewer comment.